requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.14.2
drf-yasg==1.21.7
Faker==19.2.0
filelock==3.12.2
flake8==6.0.0
gunicorn==21.2.0
identify==2.5.25
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typed-ast==1.5.5
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.1
